Fixed and Refactored DB Connection
