+++ Configs/loadenv.go
@@ -1,0 +1,15 @@
+package config
+
+import (
+	"log"
+
+	"github.com/joho/godotenv"
+)
+
+func LoadEnv() {
+	err := godotenv.Load(".env")
+
+	if err != nil {
+		log.Println("No .env file found")
+	}
+}
